lib/requirements.txt
jsonschema
pylatex
PyYAML
colorama
jinja2